+++ swarm_bot_cpp/src/__timer1__.h
@@ -10,20 +10,14 @@
 #define __TIMER1___H_
 
 
-#include <stdio.h>
-#include <asf.h>
-#include <avr/io.h>
-#include <avr/interrupt.h>
-#include <board.h>
-#include <conf_board.h>
-#include <util/delay.h>
+#include <__swarm_wold__.h>
 
 #define _TICK_US 1
 #define _TICK_MS 0.001
 
 
 void _timer1_init(void);
-void _timer1_init_prescaler(uint8_t prescaler);
+void _timer1_init_prescaler(int prescaler);
 void _timer1_stop_and_clear(void);
 void _timer1_stop_and_save(void);
